great refactoring starts
