+++ src/App.jsx
@@ -1,4 +1,4 @@
-import './App.css';
+// import './App.css';
 import Cookies from 'js-cookie';
 import { useDispatch, useSelector } from 'react-redux';
 import { setToken } from 'store/UISlice';
