+++ src/components/Search/styles.js
@@ -10,8 +10,9 @@ export const Search = styled.div `
 export const SearchSvg = styled.svg `
   width: 17px;
   height: 18px;
-  color: white;
   padding: 10px 15px 10px 8px;
+  stroke: white;
+  fill: transparent;
 `;
 
 export const SearchInput = styled.input `
