+++ src/components/index.js
@@ -1,0 +1,5 @@
+// const { Header__Sidebar } = require("./Header__Sidebar/Header__Sidebar");
+
+export { Header__Sidebar } from './Header__Sidebar/Header__Sidebar';
+export { Search } from './Search/Search';
+export { Header } from './Header/Header';
