+++ src/index.css
@@ -9,7 +9,7 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  background-color: #000;
+  background-color: #181818;
   min-height: 100vh;
   min-width: 100%;
   max-width: 100%;
