+++ src/pages/MainPage/components/styles.js
@@ -9,7 +9,7 @@ export const wrapper = styled.div`
   grid-template-columns: 15% 60% 25%;
   grid-template-rows: 70px 1fr;
   margin: 0 auto;
-  padding: 36px;
+  padding: 20px 36px;
   color: #fff;
   padding-left: 1%;
   padding-right: 1%;
